+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [1.0.3] - [2025-03-27]
+## [1.0.3] - [2025-04-06]
 
 ### Added
 - additional functionality for New-AMCard
@@ -6,9 +6,17 @@
 ### Fixed
 - New-AMTimeInput - Added warning "This function is not supported in all outlook clients. Currently only supported in Outlook on the web."
 - New-AMCard - Fixed versions to use for improved reliability in diffrent outlook clients
+- ConvertFrom-AMCard Originator ID missing from card
+- ConvertFrom-AMCard image missing from card
+- ConvertFrom-AMCard factset missing from card
+- ConvertFrom-AMCard added generated id to New-AMColumnSet
+- Add-AMElement added new function "Find-AMContainer" for better consistency in managing containers recursively.
 
 ### Feature
-- add Style parameter to New-AMImage function to allow to set avatar default = default
+- New-AMImage Style parameter to function to allow to set avatar default = default
+- New-AMContainer more options for parameter Padding valid options is now "None", "Small", "Default", "Medium", "Large", "ExtraLarge", "Custom"
+- New-AmContainer parameter option Custom in $Padding and added $CustomPadding options example @{ top = "None"; bottom = "Large"; left = "Small"; right = "Small"}
+- New-AMExecuteAction IsPrimary parameter to function to allow to set primary positive style on button
 
 
 ## [1.0.2] - [2025-03-24]